Adiciona UserValidator e ValidationException

+++ Services/User/UserService.cs
@@ -9,33 +9,39 @@ namespace Services;
 public class UserService : IService<User>
 {
     private readonly IUnitOfWork _uow;
-
     private readonly AuthData _auth;
+    private readonly FluentValidation.IValidator<User> _validator;
 
-    public UserService(IUnitOfWork uow, AuthData auth)
+    public UserService(IUnitOfWork uow, AuthData auth, FluentValidation.IValidator<User> validator)
     {
         this._uow = uow;
         this._auth = auth;
+        this._validator = validator;
     }
 
     public List<User> FindAll()
     {
         List<User> users = _uow.UserRepository.FindAll().ToList();
         if (users.Count == 0)
-            throw new Exception("Nenhum dado encontrado.");
+            throw new NotFoundException("Nenhum dado encontrado");
 
         return users;
     }
 
     public User Add(User user)
     {
+        this.Validate(user);
         user.Password = Crypter.MD5.Crypt(user.Password);
         return _uow.UserRepository.Create(user);
     }
 
     public User GetById(long id)
     {
-        return _uow.UserRepository.FindById(id);
+        User user = _uow.UserRepository.FindById(id);
+        if(user == null)
+            throw new NotFoundException("Usuário não encontrado");
+
+        return user;
     }
 
     public User Update(User user)
@@ -43,6 +49,8 @@ public class UserService : IService<User>
         User loggedInUser = _auth.LoggedInUser;
         if(loggedInUser.Id != user.Id)
             throw new AccessDeniedException("Você não tem permissão para atualizar este usuário");
+
+        this.Validate(user);
 
         if(!user.Password.Equals(loggedInUser.Password))
             user.Password = Crypter.MD5.Crypt(user.Password);
@@ -52,6 +60,19 @@ public class UserService : IService<User>
 
     public void Delete(long id)
     {
+        User loggedInUser = _auth.LoggedInUser;
+        if(loggedInUser.Id != id)
+            throw new AccessDeniedException("Você não tem permissão para remover este usuário");
+    
         _uow.UserRepository.DeleteById(id);
+    }
+
+    public void Validate(User user)
+    {
+        var validation = _validator.Validate(user);
+        if(!validation.IsValid)
+            throw new ValidationException(validation.ToDictionary());
+
+        // TODO validar se já existe Username e Email iguais
     }
 }
